Don't apply transactions of abort out-of-order.
We've started to use abort for more than just aborting (ie.
we abort for certain transitions we consider "hidden"). As a
result, we actually do need to maintain it's apply ordering
with other transitions. So, apply it's startT and finishT
at the "normal" times (ie. when it is taken off ready queue.

Bug: 283887866
Test: observe winscope in ctsverifier test from bug.
Change-Id: Ie063de29ea7a81e5874bfe7c35b29665170b362a