Do not wait for Activity#onDestroy if the activity has no process
The activity process was killed by force stopping the app package in
#teardown, but the activity was deferred to be removed from wm hierarchy
due to in transition. Before the transition completed, a kill-task-process
message was scheduled because the app task was requested to be removed
by the test. The app process that started afterward could then be
killed once the timeout occurs.

Bug: 286349050
Test: atest DisplayTests
Change-Id: I8c70c25342fd373c812f535bd2bc47da4038af04