Avoid 60ms QS inflations on UI thread after density changes
Before the change, density was always considered as "interesting" change
from InterestingConfigChanges.

After, it depends whether the flag CONFIG_DENSITY is provided or not.
To keep the default behaviour as before, the flag is kept in the default
set.

Note that it's unknown why density was handled differently from other
config fields instead of just using CONFIG_DENSITY.

Bug: 286511842
Test: adb shell wm density 300 during a perfetto trace: 60ms main thread
inflations from QS fragment are gone.

Change-Id: I7aed4dda853af80981fcaa7c80f6afff64a0c60f