Transform displayMetrics if needed
Display#getRotation returns ROTATION_0 for displays that are not
active. Display#getRealMetrics relies on rotation to flip width/height
when the display is rotated. This results in incorrect metrics being
reported from getRearDisplayMetrics() when the rear display is off
and the device is rotated.

This CL adjusts the rear display metrics so that they are reflective
of the actual device rotation.

Bug: 281952366
Test: atest WindowAreaComponentImplTests
Test: Run CTS tests starting in different initial orientations


Change-Id: I0b60039217f706b7829550d21c4f883778ca1f6d