Inject NPVC into ShadeControllerImpl
Now that NPVC is a singleton, it can be injected into SC instead
of being passed in by CentralSurfaces. Since NPVC must be injected
lazily into SC to avoid updating most classes that inject SC,
the NPVC setup code in SC must be moved to a CoreStartable. This
causes some Dagger issues for TV that required additional changes.
This change also creates an empty impl of SC to be used by Android
variants that don't have a shade, allowing us to make changes to
the SC interface without a Gerrit topic.

Fixes: 237661616
Fixes: 288867586
Test: manual, atest, and presubmits
Change-Id: Ia58178b37f3179ff10ac7476f78394e7874e790b