Reconnect if weaver service connection dies
For robustness reasons, reconnect to the Weaver service if it dies.
This should not be relied on, but it's easy to do and is the same thing
that LockSettingsService does for the Gatekeeper service.

Bug: 287705522
Test: Treehugger
Merged-In: Ic06c45353c6b7282e2f22eef13ea1c6e36fe5c58
Change-Id: Ic06c45353c6b7282e2f22eef13ea1c6e36fe5c58
(cherry picked from commit 5b21e30b66567881f40bd531d5d1750b1b987cc7)