Wait for bubbles to be loaded in BubbleDataRepositoryTest
I made some changes ag/23755383 where we might write to XML again
when loading bubbles to clear out anything that is no longer valid.

Testing this is causing some issues with another test that validates
XML is read / written correctly. Basically, the load is async so
it might write over data being used in another test. The fix
is to wait until the load is finished in BubbleDataRepositoryTest.

Additionally uses a non-mocked executor so that the callback actually
runs, before it wasn't so this test wasn't really testing properly.

Test: atest BubbleDataRepositoryTest BubblePersistentRepositoryTest --iterations 10
Bug: 288984801
Bug: 288730033
Change-Id: I461287ec40ca0a4e01185a108a23c4f25b401325