Only run flaky permissions test in postsubmit (frameworks/base)
Certain tests in these test modules have a flake rate higher than 0.5%,
which means they are out-of-SLO for Project Ironwood. We've marked them
as @FlakyTest (in a separate change). But, in this change, we also
update the TEST_MAPPING files so that don't run any flaky tests
(i.e., annotated with @FlakyTest) in presubmit, but do run them in
postsubmit.

Test: Verify tests do not run in presubmit, but do in postsubmit
Bug: 290240321
Change-Id: Ia679980a0d8aa7bc9266a3f624ff7279c19a46b3